Fix handling of urlencoded '+'
Previously, '%2B' was incorrectly converted to space.  The reason was
that page.js's `Context()` calls `decodeURLEncodedURIComponent()` when
initializing `pathname` and `Route.match()` calls
`decodeURLEncodedURIComponent()` again on the matched `params` after
splitting `pathname`.  Since `decodeURLEncodedURIComponent()` replaces
plus with space and the path is already decoded when calling it on the
`params`, pluses were incorrectly converted to spaces.

page's `Route.match()` calls `decodeURIComponent()` to decode `pathname`
before applying the path regex, so there is no need to decode the entire
path when initializing the `Context()`.

page.js's behavior seems odd.  But I haven't fully understood whether it
is a bug, so this fixes it by disabling page.js's
`decodeURLEncodedURIComponent()` via the config object.

The tests are updated to exercise the param matching with a few
urlencoded special characters.